update: (YT) simplify code

+++ Script/Youtube/main-request.ts
@@ -4,7 +4,7 @@ import { $ } from './lib/env'
 const requestMsg = new RequestMessage()
 
 try {
-  const bodyBytes = requestMsg.fromBinary($.request.bodyBytes).pure().toBinary()
+  const bodyBytes = requestMsg.fromBinary($.request.bodyBytes as Uint8Array).pure().toBinary()
   $.done({ bodyBytes })
 } catch (e) {
   $.log(e.toString())
